cleanup review preview app
