+++ figma/preview/src/lib/useFullscreen.ts
@@ -1,4 +1,5 @@
-import { useCallback, useEffect, useState } from 'react';
+import { useCallback, useState } from 'react';
+import { useEscapeKey } from './useEscapeKey';
 
 // In-app "fullscreen": hide the app chrome (header, panel, card) so the Figma
 // content fills the whole browser viewport. Not the browser's OS fullscreen.
@@ -9,14 +10,7 @@ export function useFullscreen() {
   const enter = useCallback(() => setIsFullscreen(true), []);
   const exit = useCallback(() => setIsFullscreen(false), []);
 
-  useEffect(() => {
-    if (!isFullscreen) return;
-    const onKey = (e: KeyboardEvent) => {
-      if (e.key === 'Escape') setIsFullscreen(false);
-    };
-    window.addEventListener('keydown', onKey);
-    return () => window.removeEventListener('keydown', onKey);
-  }, [isFullscreen]);
+  useEscapeKey(exit, isFullscreen);
 
   return { isFullscreen, enter, exit };
 }
