+++ figma/preview/src/main.tsx
@@ -3,9 +3,6 @@ import { createRoot } from 'react-dom/client';
 import App from './App';
 import './styles.css';
 
-// A new payload arrives via the URL hash; reload to re-parse cleanly.
-window.addEventListener('hashchange', () => location.reload());
-
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <App />
